// src/types/House.ts
export interface House {
    "url": string;
    "name": string;
    "region": string
    "coatOfArms": string
    "words": string
    "titles": Array<string>;
    "seats": Array<string>;
    "currentLord": string;
    "heir": string;
    "overlord": string;
    "founded": string;
    "founder": string;
    "diedOut": string;
    "ancestralWeapons": Array<string>;
    "cadetBranches": Array<string>;
    "swornMembers": Array<string>;
}